Add live migration support to LXD
This commit adds live migration support to LXD. As you can see from the diff,
there are several TODOs, notably:

* we need to transport the container config too (more robust config support
  will be arriving shortly, so I'll use that when it's here)
* we need to use protobufs for the control socket protocol so others can
  interact with us
* we need to use p.haul and other similar optimizations
* we need to figure out what to do with criu log files so that people have a
  nice debugging experience

To set up a container for migration, its configuration needs to have the
following entries:

    lxc.console = none
    lxc.tty = 0
    lxc.cgroup.devices.deny = c 5:1 rwm

Which can be acheived by using raw.lxc for now. You also (of course) need to
have a relatively recent version of criu installed. For this patch, I tested
with 1.4.

Closes #113

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>