exec: send term size first
While debugging something else, I noticed that the term size for exec is
racy: if the command happens to run before the goroutine in the client
sends the actual size of the terminal, we just get the default size for the
pty.

Instead, let's add a "ready" channel to the control socket handler, so that
clients can do whatever setup they need (right now just term size, but
potentially maybe other things) before the command is run.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>